+++ algorithms/0561_array_partition_i.cpp
@@ -25,7 +25,7 @@ class Solution
 
         // add the pairs
         int sum = 0;
-        for (int i = 0; i < static_cast<int>(nums.size()); i += 2)
+        for (size_t i = 0; i < nums.size(); i += 2)
         {
             sum += nums[i];
         }
